RagCode/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
faiss-cpu
sentence-transformers
python-dotenv
pypdf
python-docx